requirements.txt
# Langchain & LangGraph stack
langchain>=0.1.16
langgraph>=0.0.34
langchain-community>=0.0.28
langchain_together>=0.3.0
langchain_openai>=0.3.27
langsmith>=0.4.4

# Streamlit app
streamlit>=1.33.0

# LLM & Tokenization
openai>=1.23.0
tiktoken>=0.9.0

# PDF & file handling
pymupdf==1.23.22         # for `fitz` module
pandas>=2.2.0,<2.3.0

openpyxl                 # for Excel export

# MongoDB
pymongo>=4.13.0
dnspython                # for MongoDB URI parsing

# Environment variables
python-dotenv

# Networking & async
aiohttp>=3.9.5
httpx>=0.28.1
httpx-sse>=0.4.1

# Serialization & validation
pydantic>=2.7.1
dataclasses-json>=0.6.7

# Numerical computing
numpy>=2.3.1
